docs: explain SecureStore's `requireAuthentication` (#29981)
# Why

closes https://github.com/expo/expo/issues/29637

# How

It explains the `requireAuthentication` more. On iOS, when
`requireAuthentication` is provided, if an entry doesn't exist for the
given key, it's created with the
[kSecAttrAccessControl](https://developer.apple.com/documentation/security/ksecattraccesscontrol)
option set. No user interaction is required at this point since we're
just creating a new entry.
However, when reading or updating the value later, user is required to
authenticate using biometry.

This is different on Android where bimetric auth is required even for
adding a new value.

# Test Plan

not needed

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>