refactor(secure-store): check SecAccessControlCreateWithFlags return value (#29983)
# Why

Follow-up on https://github.com/expo/expo/pull/29981

The value returned by `SecAccessControlCreateWithFlags` is an optional,
and we should check that it's not `nil` to make sure it was created
successfully before we use it further. In practice, the `nil` value
probably almost never occurs but it's best to cover this.

# How

using `guard let` assignment

# Test Plan

You can trigger the error by providing an invalid `protection`
parameter, such as with

`SecAccessControlCreateWithFlags(kCFAllocatorDefault, "Hello there" as
CFString, .biometryCurrentSet, &error)`

I tested this on a real device. The resulting error contains a code
(that I'm returning to JS), and also a message. I could return the
message which might be more helpful but I thought in case it contains
something sensitive it's best stick with the code only. The error should
be extremely rare. Happy to change this though.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).