feat(cli): automatically set app IDs on prebuild (#30023)
# Why

Prompting for the bundle identifier and package name is a relic from
over 4-5 years ago where it was hard to change these values after
running `exp eject`. Now you can simply re-run `npx expo prebuild` to
update them whenever.

Similarly, we don't validate or manually prompt users to choose the
project name before generating a bunch of files with the default value
that Expo CLI infers from the project.

Currently we only prompt for the IDs when none can be found in the
project already, if we can guess at a generally unique value then we
recommend it as the default ('mash the enter key' flow).


<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How


This new change will perform validation on the default identifier and
check against the store for any existing app with the same identifier,
if all checks pass then we'll attempt to write the bundle identifier to
the project (this will fail if the project is using a dynamic Expo
config like app.config.js).

> A possible future iteration of this functionality could lazily infer
the app identifiers during prebuild so as to remove the constraint on
needing a static config.

If any of the validation fails, then we'll fallback to the original
behavior of prompting for an application identifier with no default
value.

- resolve ENG-12484
- I've also reworked the git messages that show when you use `--clean`
to better reflect our confidence with how this command will execute.
- Reworked the default value system to sanitize according to the package
name selector UI in Android Studio and the bundle identifier UI in
Xcode. This should greatly reduce the chances of having a prompt.

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

Manual testing of different workflows:
- `rm -rf ios app.json android; nexpo prebuild --no-install --clean`
  - prompts about uncommited changes.
  - no prompts for app IDs
- force the suggested ID to be `com.evanbacon.pillarvalley` and run `rm
-rf ios app.json android; nexpo prebuild --no-install --clean` -> warn
about name in use.
- continue -> no -> prompt as usual. Any valid name for one platform
will automatically be used for the other.
- continue -> yes -> use invalid bundle identifier anyways (this is
useful for when you're the owner of the already registered app ID, and
migrating to Expo).
- force the suggested ID to be an invalid package name:
`com.evan-bacon.pillarvalley`
  - silently prompts with no default value.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>