[core][av] fix error on new architecture mode (#30030)
# Why

fixes #29996

# How

it's a regression from #28829

- [core] fix `addUIBlock` and `executeUIBlock` crash on new architecture
mode. we don't have to access content subview after my changes.
- [av] Video is broken because the legacy `EXVideoView` is not an
ExpoFabricView and doesn't receive prop updates. it was working because
we had the view wrapper. this pr tries to add an `ExpoVideoView` to
contain the subview. that is to simulate the working state before my
change