chore(cli): add robots e2e export test (#30049)
# Why

Add a test to cover running export with a robot user enabled.

# How

Unlike a standard test where we use a private key, this key is public to
allow for testing the same locally and in the cloud and across all PRs
(not just from the team). The key is view-only for a burner account and
can be easily rotated.

# Test Plan

Was able to repro the hanging process by moving the line where we
connect the dev session. This test should hopefully catch instances in
the future when a process hangs and the export cannot close
successfully.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>