fix: swift Enumerables did not convert to a JS value (#30191)
# Why

I need this for https://github.com/expo/expo/pull/30102

When defining a swift `Record` with en `Enumberable`, the enum was not
correctly converted to a JS value. See `works back and forth with an
enum` test case which would previously fail.

In Kotlin, this already worked but I expanded the test suite to cover
this case as well.

Also added a test case `primary constructor default values are ignored
if using the 'shorthand' class syntax` that now passes but ideally
should fail. Not sure how to fix this one because we'd need to be able
to match the JS parameters to the constructor parameters to be able to
fix this - but order matters here. I'll discuss with Lukasz.

# How

expanded `packages/expo-modules-core/ios/Core/Conversions.swift` to
account for enums

# Test Plan

tests pass

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>