Fix .cjs support in @expo/server API functions
There was an error in the logic to choose between using `require(…)` or `import(…)` for an API file. The regex was checking for `.cs` or `.js` files and would `require(…)` as a result, and for anything else it would `import(…)`.

I think the intention here was to look for `.cjs` or `.js` instead.

I ran into this problem while looking into swapping my expo app to using `{"type": "module"}` in its `package.json`.