test(cli): add monorepo e2e test based on a custom fixture (#29735)
# Why

This is a draft to add a specific monorepo test to the exporting e2e
tests.

> Note, I pulled out the package linking logic. As discussed, we are
probably going for the tarball option as it's closer to how packages are
being published and keeps things better isolated. But I'll do that in a
follow-up PR.

# How

- We need a custom fixture for this, so I added one based on
`e2e/fixtures/with-router`
- Add a single test case that checks if both projects are not leaking
code, based on static routes

# Test Plan

- Test without the fix is confirmed failing in:
https://github.com/expo/expo/actions/runs/9499559787/job/26180698726?pr=29735#step:10:115
- When rebasing on the fix from #29733, it should work

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).