feat(metro): Add optimize graph mode (#30417)
# Why

- Split out of the tree shaking PR
https://github.com/expo/expo/pull/30111
- This PR contains most of the tree shaking PR but without any of the
tree shaking algorithm or docs.
- Contains all of the infra required to optimize the bundle after the
entire graph has been computed, then reconcile the bundling. This is
particularly useful to split up because most of the possible problems
will come from this part and not the actual graph crawling algorithm.

# How


- Provide a new transform setting `optimize` which preserves
import/export syntax and skips wrapping modules in an IIFE. This will
also store a reference to the ast, and the missing steps (`reconcile`)
so the transformation can be reconciled in the serializer.
- Side-effects are looked up lazily and used to create inline requires
which respect global mutations. In the future we can add upstream Metro
support for doing this in the resolver with the in-memory file map. This
will be good for caching and development workflows—neither of which are
super important here.
- The collect-dependencies module has been updated to support crawling
without mutating the AST, this is useful for sharing the import crawling
syntax with the tree shaking code later.
- This feature is hidden behind an experimental flag
`EXPO_UNSTABLE_METRO_OPTIMIZE_GRAPH=1` which will be required for tree
shaking. This flag can only be enabled when `experimentalImportSupport`
is on.

The current approach of splitting the transformation pipeline across the
bundling lifecycle will increase the chance of inconsistencies between
optimized and non-optimized bundles. This is a bit of a shortcoming with
Metro's architecture. As a result, the optimize process is much slower
and cannot be cached. This shouldn't be much of an issue though as we
only run optimization steps in production, which is often uncached (CI,
embed bundles force reset the cache).

# Test Plan

- Added E2E bundling and web test for production bundles.
- Added a number of unit tests for ensuring the re-parsing and
reconciling doesn't cause issues.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>