[doctor] Add experimental React Native Directory package check (#30496)
# Why

It is useful for us to be able to surface information about which
libraries are unmaintained and which do not support the New
Architecture, and we keep this information on React Native Directory
currently. Not sure exposing it as a default check is right, so I put it
behind a flag for now.

<img width="1138" alt="image"
src="https://github.com/user-attachments/assets/e0ff96cd-5783-46c4-81f9-d004912451ac">

# How

I added an endpoint to reactnative.directory that takes a list of
packages and returns some useful metadata about them. I then hit that
endpoint in doctor and show results based on that.

# Test Plan

I ran it in my project and it produced the expected results.

---------

Co-authored-by: Bartosz Kaszubowski <gosimek@gmail.com>