feat(metro, cli): add tree shaking (#30111)
# Why

> Most of the original PR has been landed in
https://github.com/expo/expo/pull/30417

- Metro bundler does not have any form of graph-wide tree shaking
built-in. This can lead to bundle sizes that are larger than they need
to be, especially in the case of icon libraries.
- This PR (which is partially an RFC) aims to add a system for removing
unused imports and exports from a production bundle.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Leverage the new "optimize graph" mode to perform holistic
optimizations in the serializer. https://github.com/expo/expo/pull/30417
- Tree shaking can be enabled with the new env var
`EXPO_UNSTABLE_TREE_SHAKING=1`.
- This PR also adds a "barrel expansion" system which attempts to
replace `export * from '...'` statements with `export { View, Image, etc
} from '...'` statements to enable more graph reduction.
- Empty modules (containing only comments and directives) are removed
from the bundle.
- Side-effecty imports (imports with zero specifiers, `import "foo"` and
`import {} from "foo"`) are preserved unless marked otherwise by a
`package.json`'s `sideEffects` field.
- Custom imports such as `require.context` and `require.resolveWeak` are
left in place.
- Removed exports with source will trigger a recrawl to ensure any new
unused imports/exports are recursively removed.
- The feature is artificially forced to only run when
`EXPO_UNSTABLE_TREE_SHAKING=1` is set, and when bundling for production.
Tree shaking is disabled for SSR bundles. I may expand this to be
web-only too just due to a lack of continuous testing for native
runtimes.
- Added some external fixes
https://github.com/software-mansion/react-native-reanimated/pull/6171
https://github.com/expo/expo/pull/30010
https://github.com/expo/expo/pull/29980
https://github.com/expo/expo/pull/29964
https://github.com/expo/expo/pull/29963
https://github.com/software-mansion/react-native-gesture-handler/pull/2972

# Further work

Some ideas that I've had but aren't implemented in this PR:

- Use the worker farm to parallelize transformation in the serializer.
This could be tricky because we'd need to convert the AST to a serial
format then re-parse it back.
- ~~Fork collectDependencies to fully track the import/exports so we can
reuse the same import crawling code and references. This would also
speed up the bundling as we could parallelize and cache the pass.~~
https://github.com/expo/expo/pull/30140
- Use a totally custom syntax for matching import/exports, e.g.
`_$$_IMPORT_DEFAULT, _$$_IMPORT_ALL` (and some new export equivalents)
to reduce the number of transforms we run in the serializer. All
import/export's and the iife wrap could be run at transform-time and
we'd simply be mutating a valid JS bundle in the serializer. This would
be very challenging though and make it harder for users to contribute to
the implementation.
- ~~Upstream the hack on transformSyncRequire to support require.context
with one parameter. https://github.com/facebook/metro/pull/1129~~
- Detect if a module is CJS according to Node.js heuristic, e.g. `.cjs`,
package exports, etc. then opt it out of tree shaking.
- [ ] update source maps

# Test Plan

- Added an e2e tree shaking test for web. 
- Added mini-metro tests to account for the known and expected tree
shaking behaviors.
- Added mini-metro tests for side-effect handling (need more here).
- Ensured all apps in expo/apps can be transformed and run with tree
shaking enabled.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>