feat(jest-expo): add experimental RSC testing library (#29404)
# Why

- Many libraries will need at least minor changes to become compatible
with the new universal React Server Components feature we're adding to
Expo Router.
- This PR aims to simplify the migration by splitting out part of
feature and adding it to `jest-expo` (depends on the existing support
already landed in the CLI).
- ~~Unfortunately, this won't be mergable for now since it requires
React 19, and the standard testing library we're using depends on React
18.~~ Using the new react and react-dom vendor packages in expo/cli, we
can land this without setting resolutions in the repo.
- ~~My plan is to use this branch as a testing ground, then move SDK
packages into the upstream with basic RSC support and no tests.~~ I've
reverted all non-important changes so we can try to land this and get
RSC in the CI.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Add a new preset `jest-expo/rsc` for bundling tests in React Server
mode.
- Also has platform-specific variants `jest-expo/rsc/ios`,
`jest-expo/rsc/android`, `jest-expo/rsc/web`.
- Add a new matcher `await expect(<jsx />).toMatchFlight(...)` which
compiles the input node to a pseudo-RSC payload that can be used to
generally inspect client boundaries. This is based loosely on the one in
Expo Router but outputs a more CI-friendly object.
- To prevent test projects from ingesting each-other, I've added
`__rsc_tests/` which is the only place where RSC tests will be matchable
from. This has also been added to the npmignore of expo module scripts.

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- `expo-linear-gradient` has been migrated over to support RSC.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>
Co-authored-by: James Ide <ide@users.noreply.github.com>