[expo-notifications][Android] map Expo and Firebase notifications correctly (#30615)
Since #7280, some fields in Android notification content have been remapped. This is done in order to align the results seen in JS with the fields passed into the Expo push notification service when it sends notifications, and to align the API across the iOS and Android platforms.

However, this mapping is also happening when notifications are sent directly from Firebase. This leads to unexpected results where the `data` part of the notification payload does not appear correctly in the JS notification structures.

This PR fixes that by heuristically detecting whether a notification is sent from Expo or not, and performing the appropriate mapping in either case.

- Modified code in `NotificationSerializer` to detect when the `body` attribute is a JSON string (indicating that the notification was sent from Expo), and remapping the resulting bundles as appropriate for both cases.
- Modified `NotificationsEmitter` and `mapNotificationResponse` in the JS code to correctly unpack data strings in the case of Expo notifications.
- Added extensive debug logging, wrapped by `BuildConfig.DEBUG` so the logging does not happen in release/production builds.

- Both Firebase and Expo notifications should be mapped correctly in the test app
- Notification responses should be mapped correctly in both cases when tapping on a notification while the app is in background or not running
- CI should pass

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).