fix(expo-updates): pass absolute entry path from gradle plugin and convert to relative path in Node (#30654)
# Why

This is a follow-up of #30652, #30633, #30621

Kinda annoying that we use relative entry paths for iOS, and absolute for Android. But it is what it is.

# How

This moves the incorrect relative path modification in the `expo-updates-gradle-plugin` to the Node script itself, where we can reuse the functions from `@expo/config/paths`. These shared functions takes the server or project root into account.

# Test Plan

See CI, specifically test-suite android.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).