[autolinking] introduce aar files autolinking as gradle projects  (#30706)
# Why

for expo modules to wrap existing aar. it could be a solution for #27985
close ENG-12425

# How

`expo-module.config.json` can now specify an aar file and it could be created as a gradle project. For example, if an expo-hello project having the config,

```json
{
  "gradleAarProjects": [
    {
      "name":  "test-aar",
      "aarFilePath": "android/libs/test.aar"
    }
  ]
}
```

autolinking will create a project named `expo-hello$test-aar`. the main project could then add the dependency:

```gradle
dependencies {
  implementation project(":expo-hello\$test-aar")
  // Or even without hardcoded main project name
  implementation project(":${project.name}\$test-aar")
}
```