[expo-notifications][Android] Eliminate unsupported types when processing notification intents from onCreate/onNewIntent (#30750)
On Android, when a notification is tapped to open the app, the intent
passed into the Android lifecycle listeners may have properties in the
`extras` bundle that are not supported in converting to `WritableMap`
for sending events to JS. This potentially leads to a crash.

This PR fixes #30738.

- New methods in `Utils` to filter out bundle properties that are not
convertible.
- `NotificationSerializer` is modified to use the above new methods to
filter out unconvertible values when processing the intent extras in
`onCreate()` and `onNewIntent()` from
`ExpoNotificationLifecycleListener`.

- CI should pass
- The test app should not crash when this condition is encountered

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).