[expo-font]: add getLoadedFonts() function (#30431)
# Why

This is part of the effort to make is easier to check (1) if font is
loaded (2) load it asynchronously. The goal is to expose this
information so that it can be consumed easily from JavaScript which
could live in community libs without depending on Expo, such as RN
vector icons.


# How

Added a new `getLoadedFonts()` function to native modules. This
synchronously gives the set of the fonts that were either statically
bundled or loaded dynamically at runtime.

# Test Plan

existing tests pass. Also added some tests for the new function as well
as a fix for an issue on web (see `works for FontSource object where uri
is a module id` test case).

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Will Schurman <wschurman@expo.io>
Co-authored-by: Tomasz Sapeta <tomasz.sapeta@swmansion.com>