docs: regenerate api docs json files (#30953)
# Why

working on some other changes to how docs are generated and I wanted to start with a "clean state". 

# How

These changes are the result of running "et gdad" for all packages.

Perhaps there should be a CI check for this?

# Test Plan

docs look okay

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).