docs: support generating and rendering TypeOperator target types (#30954)
# Why

Some docs use types that use keywords like `readonly` or `keyof`. Our docs currently don't play well with these. This PR attempts to fix it.

The solution is (1) keep more data in `json` files generated by `et gdad` and (2) use that information in the renderer

# How

The json output that's of interest to us can look like this:

```json
{
    "type": "typeOperator",
    "operator": "keyof",
    "target":
    {
        "type": "reference",
        "name": "VideoPlayerEvents",
        "package": "expo-video"
    }
}
```

but `target` can also be a number. I'm not sure what all cases there are but I took an approach that should "filter" only those that are of interest to us.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff. This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).