[create-expo] add a comment to sync the forbidden names list in WWW (#31085)
# Why

I copied over the forbidden names validation logic to WWW to prevent issues when creating a project from the website, so I wanted to add a comment here to tell authors to update that list in WWW after changing it here.

The WWW PR number for reference: #16354 

# How

I added a comment above the `FORBIDDEN_NAMES` list.

# Test Plan

Nothing should change. Inspect the diff to read the comment.