feat(babel-preset-expo): Add basic support for using DOM Components in RSC projects (#31080)
# Why

Projects bundled with RSC should be able to import and render DOM
components in some basic capacity. The more advanced uses (passing
server actions or RSC as props) will need to be explored further in the
future.

# How

Treat `"use dom"` the same as `"use client"` when bundling for
react-server envs. This will effectively opt the DOM component out of
SSR and require the app to recurse back to the server for it.

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

Added unit test for basic behavior

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>