[android][image] Fix NoSuchMethodException on 0.75 and proguard (#31153)
# Why
Closes #31118

# How
On 0.75 `ReactViewBackgroundDrawable` has been deprecated and now
extends `CSSBackgroundDrawable`. When proguard is enabled this class is
obfuscated so reflection fails to find the method. Instead of adding
rules for both of these classes, we can just add one for
`CSSBackgroundDrawable` and use that instead on 0.75. This class is
found through reflection because it does not exist in 0.74

# Test Plan
Tested in a new project in release mode using 0.74 and 0.75 with
proguard enabled.

---------

Co-authored-by: Brent Vatne <brentvatne@gmail.com>