chore(router): refactor hydration setting (#31267)
# Why

- We were completely invalidating the cache to change the rendering
output mode. This was mostly just an artifact of when expo-router was in
a different repo.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- I've refactored the hydration marker to not invalidate the cache at
all or even go through Babel, we just inject it as a script global in
the HTML during static rendering.
- This is pulled out of the RSC branch.
- I did notice however that we aren't enabling hydration in development
at all. It appears to be broken on the suspense boundaries. We might not
be able to fix this until we have RSC support enabled.

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

- E2E test for this will pass.
- Dev works as expected.
- Cache doesn't invalidate between output modes.


<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).