[image-picker][ios] Add support for picking live photos (#30957)
# Why

We need a way do allow users to get live photos for `expo-live-photo`
when it's created.

# How

Added support for picking live photos from the library. This is fully
backwards compatible, because `useLivePhotos` property is false by
default.

Live photo is returned in two parts - the image asset and a nested video
asset. The image and video can be combined natively into a `PHLivePhoto`
with
[`request(withResourceFileURLs:placeholderImage:targetSize:contentMode:resultHandler:)`](https://developer.apple.com/documentation/photokit/phlivephoto/1616434-request)

We need to preserve the exif data for the function to work, that's why
we can't crop or compress the image and we have to use a separate
function for handling livePhotos.

# Test Plan

Tested in Simulator and a physical device in BareExpo