[expo-video] Add support for playback from a `require` result, add usage section to the docs (#30837)
# Why

Fixes: https://github.com/expo/expo/issues/30789

Some users want to play assets from the `assets` directory using the
`require` function.

# How

Used the `resolveAssetSource` from `react-native` to get the uri of the
local asset.
The `require` can be either used as the video source or uri of the
source (for example when someone wants to override the source metadata)

# Test Plan

Tested the example in BareExpo in both Release and Debug modes on
Android and iOS

(cherry picked from commit 0cc07241c47209085f83be8b12bfb3564de03344)