[ios][expo-store-review] Add migration for SkStoreReviewController on iOS 16+
closes #[ENG-12917](https://linear.app/expo/issue/ENG-12917/migrate-away-from-deprecated-skstorereviewcontroller-in-expo-store)

#Why
SkStoreReviewController is marked as deprecated in the docs and will be removed after iOS 18: (https://developer.apple.com/documentation/storekit/skstorereviewcontroller).

#How
This commit fixes this by adding a new availability check that uses `Appstore.requestReview(in:)` method instead on iOS 16 and up. (https://developer.apple.com/documentation/storekit/appstore/3954432-requestreview)

In addition we're now using the new Swift 5.5 @MainActor notation to ensure that our methods run on the main thread. This is required for the `Appstore.requestReview(in:)` method to compile successfully.

Task/@MainActor is available from iOS 13.
https://developer.apple.com/documentation/swift/mainactor/
https://developer.apple.com/documentation/swift/task

Also added a new private function to get the active foreground scene in the module to avoid code duplication.

# Test Plan
expo-go runs without error on iOS.
Should be tested on multiple platforms like 13, 14 and 16 and above.