[video][docs] Document and unify the preload behavior across platforms (#30626)
# Why

It is currently possible to preload videos on iOS, but not on Android,
as the player will start preloading the data only after it has been
mounted in a VideoView.

Fixes https://github.com/expo/expo/issues/30683

# How

Made Android load video data even when not mounted in a VideoView.
If you think that the default behavior should be that the data is not
loaded and a function/property has to be called/changed to start
preloading let me know below.

Added a docs section about how preloading can be used.

# Test Plan

Tested in BareExpo on iOS and Android

(cherry picked from commit 8f420ee29dd2d6db77abe92fe43cfe26af626b8b)