Added warning for background location when running Expo Go
Background location services (including geofencing) does not work in Expo Go (it is available in the Simulator on iOS) so we need to emit a warning when users are using background locations in ExpoGo with some explanation about development builds.

This commit adds a warning that will be emitted on the console if running Expo Go client. The warning tries to clearify the above issue.

Relates to #ENG-12301