Prevent errors if podspecPath is undefined (#20789)
I ran into a case where the podspecPath was undefined while running yarn prepare. Rather than throwing an error because you can't split something that is undefined, we should simply set it as an empty string. This is assuming we shouldn't instead catch the error and present a meaningful resolution.

# Why

To prevent errors if the podspecPath is undefined.

# How

It was preventing yarn prepare from running successfully and otherwise doesn't seem to have adverse effects.

# Test Plan

It would be good to test with other plugins running yarn prepare, but it's a small change that shouldn't have adverse effects unless there is a reason we should be catching if the podspecPath is not defined. If so, we should instead throw a meaningful error.